feat(env-check): fix tests
